pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "research"
version = "0.1.0"
description = "Research paper analysis and RAG system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.3.1",
    "braintrust>=0.3.9",
    "bs4>=0.0.2",
    "dspy>=3.0.4",
    "fastapi>=0.115.0",
    "fastembed>=0.7.3",
    "mem0ai>=1.0.1",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.16.0",
    "requests>=2.32.0",
    "supabase>=2.24.0",
    "uvicorn>=0.32.0",
]

[tool.hatch.build.targets.wheel]
packages = ["arxiv_parser", "database", "memory", "rag"]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
]
